third_party/WebKit/LayoutTests/http/tests/inspector/websocket/timeline-websocket-event.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../timeline-test.js"></script>
<script>

function performActions(callback)
{
    var ws = new WebSocket("ws://127.0.0.1:8880/simple");
    ws.onclose = callback;
}

function test()
{
    InspectorTest.invokeAsyncWithTimeline("performActions", finish);

    function finish()
    {
        InspectorTest.printTimelineRecords("WebSocketCreate");
        InspectorTest.printTimelineRecords("WebSocketSendHandshakeRequest");
        InspectorTest.printTimelineRecords("WebSocketReceiveHandshakeResponse");
        InspectorTest.printTimelineRecords("WebSocketDestroy");
        InspectorTest.completeTest();
    }
}

if (!window.testRunner)
    setTimeout(performActions, 2000);

</script>
</head>

<body onload="runTest()">
<p>
Tests the Timeline events for WebSocket
</p>

</body>
</html>
